requirements.txt
selenium
fake_useragent
python-dotenv